+++ examples/simple/src/Models.js
@@ -29,7 +29,7 @@ export class MyModelRoot extends ModelRoot {
 
     init(...args) {
         super.init(...args);
-        console.log("Start root model!!!");
+        console.log("Start root model!");
         console.log(Constants);
 
         this.base = BaseActor.create({pawn: "GroundPawn"});