+++ packages/kernel/CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Pawn destruction handled by PawnManager, not Pawn itself
 - Services no longer have start-up options
 - Pawns don't update their draw tranform if they don't move
+- PawnManager can be subclassed
 
 ### Fixed
 - m4_getRotation() correctly ignores scale